Wrote tests for Game Generator, and fixed some bugs some related classes

+++ main/Entities/InterviewIntern.java
@@ -13,13 +13,12 @@ public class InterviewIntern extends Intern{
     /**
      * Construct an interviewIntern, giving them the given interviewInternName,
      * interviewInternAge and interviewInternSkills.
-     *
-     * @param interviewInternName   The name of this Entities.Intern
+     *  @param interviewInternName   The name of this Entities.Intern
      * @param interviewInternAge    The age of this Entities.Intern
      * @param interviewInternSkills A map of skills where the key is the name of the skill and the value is the percentage.
      */
     public InterviewIntern(String interviewInternName, int interviewInternAge,
-                           HashMap<String, Integer> interviewInternSkills){
+                           HashMap<String, Double> interviewInternSkills){
         super(interviewInternName, interviewInternAge, interviewInternSkills);
     }
